\name{checkDataIntegrity-methods}
\docType{methods}
\alias{checkDataIntegrity}
\alias{checkDataIntegrity-methods}
\alias{checkDataIntegrity,data.frame,data.frame-method}

\title{Methods for Function checkDataIntegrity in Package `COSTcore'}
\description{
Methods to check that all data in the "current" data frame have related records in the "target" data frame. These functions are for data integrity checking when validating csData objects and are not intended to be called directly by users. }
\section{Methods}{
\describe{
\item{checkDataIntegrity}{\code{signature(target = "data.frame", current = "data.frame")}: Checks that data in the "current" data frame have related records in the "target" data frame.}
}}
\keyword{methods}
