\name{check-methods}
\docType{methods}
\alias{checkNms-methods}
\alias{checkNms,data.frame,character-method}
\alias{checkTRnms-methods}
\alias{checkTRnms,data.frame-method}
\alias{checkHHnms-methods}
\alias{checkHHnms,data.frame-method}
\alias{checkSLnms-methods}
\alias{checkSLnms,data.frame-method}
\alias{checkHLnms-methods}
\alias{checkHLnms,data.frame-method}
\alias{checkTRpk-methods}
\alias{checkTRpk,data.frame-method}
\alias{checkHHpk-methods}
\alias{checkHHpk,data.frame-method}
\alias{checkSLpk-methods}
\alias{checkSLpk,data.frame-method}
\alias{checkHLpk-methods}
\alias{checkHLpk,data.frame-method}
\alias{checkCApk-methods}
\alias{checkCApk,data.frame-method}
\title{Check methods in COSTcore}
\description{
	This are check methods used to validate S4 classes. These are meant for interbnal usage only, not to be directly called by users.
}
\section{Methods}{
\item{checkNms}{Check names in a data.frame against a character vector.}
\item{checkTRnms}{Wrapper for TR checkNms.}
\item{checkHHnms}{Wrapper for HH checkNms.}
\item{checkHLnms}{Wrapper for HL checkNms.}
\item{checkSLnms}{Wrapper for SL checkNms.}
\item{checkTRpk}{Check TR primary key.}
\item{checkHHpk}{Check HH primary key.}
\item{checkSLpk}{Check SL primary key.}
\item{checkHLpk}{Check HL primary key.}
}
\keyword{methods}
