\name{coerceCons-methods}
\docType{methods}
\alias{coerceCons}
\alias{coerceCons-methods}
\alias{coerceCons,data.frame,data.frame-method}
\title{Coerces df columns into specific classes.}
\description{Specific 'consolidated' version of 'coerceDataframeColumns' method. These functions are for internal usage only and are not intended to be called directly by users.}
\section{Methods}{
\describe{
\item{coerceCons}{\code{signature(object = "data.frame", refObject = "data.frame")}: Checks the class of \code{refObject} columns and coerces the \code{object} columns in agreement.}
}}
\keyword{methods}

