% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csDataCons-class.R
\docType{methods}
\name{csDataCons,csDataVal,strIni-method}
\alias{csDataCons,csDataVal,strIni-method}
\title{semester' field is added to ca table}
\usage{
\S4method{csDataCons}{csDataVal,strIni}(object, objStrat,
  desc = "Unknown stock", ...)
}
\description{
semester' field is added to ca table
}
