% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csDataCons-class.R
\name{csDataConsCATR}
\alias{csDataConsCATR}
\title{number of days' and 'number of FO' fields should be updated in tr table according to stratification and ca table --> TO DO}
\usage{
csDataConsCATR(object, objStrat, desc = "Unknown stock", ...)
}
\description{
number of days' and 'number of FO' fields should be updated in tr table according to stratification and ca table --> TO DO
}
