\name{rbind2-methods}
\docType{methods}
\alias{rbind2,ceData,ceData-method}
\alias{rbind2,ceDataCons,ceDataCons-method}
\alias{rbind2,clData,clData-method}
\alias{rbind2,clDataCons,clDataCons-method}
\alias{rbind2,csData,csData-method}
\alias{rbind2,csDataCons,csDataCons-method}
\title{rbind2 for `COSTcore' classes}
\description{This method implements rbind2 for the classes provided by COSTcore.}
\section{Methods}{
\describe{
\item{}{\code{signature(x = "ceData", y = "ceData")}}
\item{}{\code{signature(x = "ceDataCons", y = "ceDataCons")}}
\item{}{\code{signature(x = "clData", y = "clData")}}
\item{}{\code{signature(x = "clDataCons", y = "clDataCons")}}
\item{}{\code{signature(x = "csData", y = "csData")}}
\item{}{\code{signature(x = "csDataCons", y = "csDataCons")}}
}}
\keyword{methods}

