\name{sole.ce}
\alias{sole.ce}
\docType{data}
\title{Sole commercial effort data.}
\description{
A ceData object containing sole data to be used for testing and development purposes.  }
\usage{sole.ce}
\format{Formal class 'ceData' [package "COSTcore"] with 2 slots: \code{desc} and \code{ce}.}
\section{WARNING}{
This data is provided is for testing only and should not be used for any other purpose. 
}
\details{
To Do
}
\source{
To Do
}
\examples{
data(sole.ce)
}
\keyword{datasets}

