\name{sole.cl}
\alias{sole.cl}
\docType{data}
\title{Sole commercial landings data.}
\description{
A clData object containing sole data to be used for testing and development purposes.  
}
\usage{sole.cl}
\format{Formal class 'clData' [package "COSTcore"] with 2 slots: \code{desc} and \code{cl}.}
\section{WARNING}{
This data is provided is for testing only and should not be used for any other purpose. 
}
\details{
To Do
}
\source{
To Do
}
\examples{
data(sole.cl)
}
\keyword{datasets}
