\name{sole.cs}
\alias{sole.cs}
\docType{data}
\title{Sole CS data.}
\description{
Sole csData object used for testing and developing pourposes.
}
\usage{data(sole.cs)}
\format{
  The format is:
Formal class 'csData' [package "COSTcore"] with 6 slots
  ..@ desc: chr "my stock"
  ..@ tr  :'data.frame':	19 obs. of  15 variables:
  ..@ hh  :'data.frame':	227 obs. of  27 variables:
  ..@ sl  :'data.frame':	86 obs. of  18 variables:
  ..@ hl  :'data.frame':	156 obs. of  16 variables:
  ..@ ca  :'data.frame':	1 obs. of  27 variables:
}
\section{WARNING}{
The data provided is for testing pourposes only and can not be used for any other objective.
}
\details{
ToDo
}
\source{
ToDo
}
\examples{
data(sole.cs)
## maybe str(sole.cs) ; plot(sole.cs) ...
}
\keyword{datasets}
