\name{subset-methods}
\docType{methods}
\alias{subset-methods}
\alias{subset,ceData-method}
\alias{subset,ceDataCons-method}
\alias{subset,clData-method}
\alias{subset,clDataCons-method}
\alias{subset,csData-method}
\alias{subset,csDataCons-method}
\title{subset for `COSTcore' classes}
\description{
This method implements subseting for classes provided by COSTcore.
}
\section{Methods}{
\describe{

\item{x = "ceData"}{subseting \code{ce} slot.}

\item{x = "ceDataCons"}{subseting \code{ce} slot.}

\item{x = "clData"}{subseting \code{cl} slot.}

\item{x = "clDataCons"}{subseting \code{ce} slot.}

\item{x = "csData"}{subseting object based on \code{tr} slot.}

\item{x = "csDataCons"}{subseting object based on \code{tr} slot.}
}}
\keyword{methods}

