\name{utils-methods}
\docType{methods}
\alias{dim,ceData-method}
\alias{head,ceData-method}
\alias{summary,ceData-method}
\alias{tail,ceData-method}
\alias{dim,clData-method}
\alias{head,clData-method}
\alias{summary,clData-method}
\alias{tail,clData-method}
\alias{dim,csData-method}
\alias{head,csData-method}
\alias{summary,csData-method}
\alias{tail,csData-method}
\alias{dim,ceDataCons-method}
\alias{head,ceDataCons-method}
\alias{summary,ceDataCons-method}
\alias{tail,ceDataCons-method}
\alias{dim,clDataCons-method}
\alias{head,clDataCons-method}
\alias{summary,clDataCons-method}
\alias{tail,clDataCons-method}
\alias{dim,csDataCons-method}
\alias{head,csDataCons-method}
\alias{summary,csDataCons-method}
\alias{tail,csDataCons-method}
\title{Utils methods in COSTcore}
\description{
Utilities to be used on sQuote{COST} classes that return meta information about the object's contents.
}
\section{Methods}{
\describe{
	\item{dim}{\code{signature(object)}: dimensions.}
	\item{head}{\code{signature(object)}: first ten rows of data.}
	\item{summary}{\code{signature(object)}: summary statistics.}
	\item{tail}{\code{signature(object)}: last ten rows of data.}
}}
\seealso{
\code{\link{dim}}, \code{\link{head}}, \code{\link{summary}}, \code{\link{tail}} 
}
\keyword{methods}

