\name{checkDataIntegrity-methods}
\docType{methods}
\alias{checkDataIntegrity}
\alias{checkDataIntegrity-methods}
\alias{checkDataIntegrity,data.frame,data.frame-method}

\title{Methods for Function checkDataIntegrity in Package `COSTcore'}
\description{
Method to check that all data in current has a related record in target. It is used to check data integrity when validating csData objects.
}
\section{Methods}{
\describe{
\item{target = "data.frame", current = "data.frame"}{Check that current has related records in target.}
}}
\keyword{methods}
