\name{coerceDataFrameColumns-methods}
\docType{methods}
\alias{coerceDataFrameColumns}
\alias{coerceDataFrameColumns-methods}
\alias{coerceDataFrameColumns,data.frame,data.frame-method}
\title{Coerces data.frame columns into specific classes.}
\description{Methods for function coerceDataFrameColumns in Package `COSTcore'.Internal usage, should not be called directly by the user.
}
\section{Methods}{
\describe{
\item{\code{coerceDataFrameColumns(object = "data.frame", refObject = "data.frame")}}{Checks the class of \code{refObject} columns and coerces the \code{object} columns in agreement.}
}}
\keyword{methods}

