\name{coerceDataFrameColumns-methods}
\docType{methods}
\alias{coerceDataFrameColumns}
\alias{coerceDataFrameColumns-methods}
\alias{coerceDataFrameColumns,data.frame,data.frame-method}
\title{Coerces data.frame columns into specific classes.}
\description{Methods for the function \emph{coerceDataFrameColumns}. These functions are for internal usage only and are not intended to be called directly by users.}
\section{Methods}{
\describe{
\item{coerceDataFrameColumns}{\code{signature(object = "data.frame", refObject = "data.frame")}: Checks the class of \code{refObject} columns and coerces the \code{object} columns in agreement.}
}}
\keyword{methods}

