\name{rbind2-methods}
\docType{methods}
\alias{rbind2-methods}
\alias{rbind2,ceData,ceData-method}
\alias{rbind2,ceDataCons,ceDataCons-method}
\alias{rbind2,clData,clData-method}
\alias{rbind2,clDataCons,clDataCons-method}
\alias{rbind2,csData,csData-method}
\alias{rbind2,csDataCons,csDataCons-method}
\title{rbind2 for `COSTcore' classes}
\description{
This method implements rbind2 for classes provided by COSTcore.
}
\section{Methods}{
\describe{
\item{x = "ceData", y = "ceData"}{rbinds x and y}

\item{x = "ceDataCons", y = "ceDataCons"}{rbinds x and y}

\item{x = "clData", y = "clData"}{rbinds x and y}

\item{x = "clDataCons", y = "clDataCons"}{rbinds x and y}

\item{x = "csData", y = "csData"}{rbinds x and y}

\item{x = "csDataCons", y = "csDataCons"}{rbinds x and y}
}}
\keyword{methods}

