\name{sole.ce}
\alias{sole.ce}
\docType{data}
\title{Sole's commercial effort data.}
\description{
Sole ceData object used for testing and developing pourposes.
}
\usage{sole.ce}
\format{Formal class 'ceData' [package "COSTcore"] with 2 slots: \code{desc} and \code{ce}.}
\section{WARNING}{
The data provided is for testing pourposes only and can not be used for any other objective.
}
\details{
ToDo
}
\source{
ToDo
}
\examples{
data(sole.ce)
}
\keyword{datasets}

