\name{sole.cl}
\alias{sole.cl}
\docType{data}
\title{Sole's commercial landings data.}
\description{
Sole clData object used for testing and developing pourposes.
}
\usage{sole.cl}
\format{Formal class 'clData' [package "COSTcore"] with 2 slots: \code{desc} and \code{cl}.}
\section{WARNING}{
The data provided is for testing pourposes only and can not be used for any other objective.
}
\details{
ToDo
}
\source{
ToDo
}
\examples{
data(sole.cl)
}
\keyword{datasets}
