\name{sole.cs}
\alias{sole.cs}
\docType{data}
\title{Sole's commercial sampling data.}
\description{
Sole csData object used for testing and developing pourposes.
}
\usage{sole.cs}
\format{Formal class 'csData' [package "COSTcore"] with 6 slots: \code{desc}, \code{tr}, \code{hh}, \code{sl}, \code{hl} and \code{ca}.}
\section{WARNING}{
The data provided is for testing pourposes only and can not be used for any other objective.
}
\details{
ToDo
}
\source{
ToDo
}
\examples{
data(sole.cs)
}
\keyword{datasets}
