\name{sole.cs}
\alias{sole.cs}
\docType{data}
\title{Sole commercial sampling data.}
\description{
A csData object containing sole data to be used for testing and development purposes.  
}
\usage{sole.cs}
\format{Formal class 'csData' [package "COSTcore"] with 6 slots: \code{desc}, \code{tr}, \code{hh}, \code{sl}, \code{hl} and \code{ca}.}
\section{WARNING}{
This data is provided is for testing only and should not be used for any other purpose. 
}
\details{
To Do
}
\source{
To Do
}
\examples{
data(sole.cs)
}
\keyword{datasets}

