\name{subset-methods}
\docType{methods}
\alias{subset,ceData-method}
\alias{subset,ceDataCons-method}
\alias{subset,clData-method}
\alias{subset,clDataCons-method}
\alias{subset,csData-method}
\alias{subset,csDataCons-method}
\title{subset for `COSTcore' classes}
\description{
This method implements subseting for classes provided by COSTcore.
}
\section{Methods}{
\describe{
\item{}{\code{signature(x = "ceData")}}
\item{}{\code{signature(x = "ceDataCons")}}
\item{}{\code{signature(x = "clData")}}
\item{}{\code{signature(x = "clDataCons")}}
\item{}{\code{signature(x = "csData")}}
\item{}{\code{signature(x = "csDataCons")}}
}}
\keyword{methods}

